chore(sdk): rename hook
